Update files, Alura, Introdução a C, Aula 3.6

+++ alura/c/adivinhacao.c
@@ -47,7 +47,7 @@ int main() {
 
     tentativas++;
 
-    double pontosperdidos = (chute - numerosecreto) / 2.0;
+    double pontosperdidos = (chute - numerosecreto) / (double)2;
     pontos = pontos - pontosperdidos;
 
   }
